+++ app/src/main/java/com/octo/entourage/model/Poi.java
@@ -18,11 +18,11 @@ public class Poi {
 
     private String audience;
 
-    private long categoryId;
+    private int categoryId;
 
-    private long longitude;
+    private double longitude;
 
-    private long latitude;
+    private double latitude;
 
     public long getId() {
         return id;
@@ -88,27 +88,27 @@ public class Poi {
         this.audience = audience;
     }
 
-    public long getCategoryId() {
+    public int getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(long categoryId) {
+    public void setCategoryId(int categoryId) {
         this.categoryId = categoryId;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
